+++ src/main/java/com/fourthread/ozang/module/domain/clothes/service/TestImageService.java
@@ -1,5 +1,6 @@
 package com.fourthread.ozang.module.domain.clothes.service;
 
+import com.fourthread.ozang.module.domain.storage.ImageService;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.context.annotation.Profile;
 import org.springframework.stereotype.Service;
